move test suite outside of Lib/ufoLib
Fixes https://github.com/unified-font-object/ufoLib/issues/165
