+++ setup.cfg
@@ -40,6 +40,7 @@ license_file = LICENSE.txt
 minversion = 3.0.2
 testpaths = 
 	ufoLib
+	tests/
 doctest_optionflags = ALLOW_UNICODE ALLOW_BYTES
 addopts = 
 	-r a
