+++ tests/test_etree.py
