+++ tests/testdata/DemoFont.ufo/fontinfo.plist
