+++ tests/testdata/DemoFont.ufo/glyphs/O_.glif
