+++ tests/testdata/DemoFont.ufo/glyphs/testglyph1.reversed.glif
