+++ tests/testdata/DemoFont.ufo/metainfo.plist
