+++ tests/testdata/TestFont1 (UFO1).ufo/glyphs/contents.plist
