+++ tests/testdata/TestFont1 (UFO1).ufo/metainfo.plist
