+++ tests/testdata/TestFont1 (UFO2).ufo/fontinfo.plist
