+++ tests/testdata/TestFont1 (UFO2).ufo/glyphs/contents.plist
