+++ tests/testdata/TestFont1 (UFO2).ufo/lib.plist
