+++ tests/testdata/UFO3-Read Data.ufo/data/org.unifiedfontobject.directory/bar/lol.txt
