+++ tests/testdata/UFO3-Read Data.ufo/data/org.unifiedfontobject.directory/foo.txt
