+++ tests/testdata/UFO3-Read Data.ufo/metainfo.plist
